Пересмотрел структуру store: ROOM
