+++ project/src/store/root-reducer.js
@@ -1,25 +1,19 @@
 import {combineReducers} from 'redux';
 import {city} from './city/city.js';
-import {neighborOffers} from './neighbor-offers/neighbor-offers.js';
-import {offer} from './offer/offer.js';
 import {offers} from './offers/offers.js';
-import {reviews} from './reviews/reviews.js';
+import {room} from './room/room.js';
 import {user} from './user/user.js';
 
 export const NameSpace = {
   CITY: 'CITY',
-  NEIGHBOR_OFFERS: 'NEIGHBOR_OFFERS',
-  OFFER: 'OFFER',
   OFFERS: 'OFFERS',
-  REVIEWS: 'REVIEWS',
+  ROOM: 'ROOM',
   USER: 'USER',
 };
 
 export default combineReducers({
   [NameSpace.CITY]: city,
-  [NameSpace.NEIGHBOR_OFFERS]: neighborOffers,
-  [NameSpace.OFFER]: offer,
   [NameSpace.OFFERS]: offers,
-  [NameSpace.REVIEWS]: reviews,
+  [NameSpace.ROOM]: room,
   [NameSpace.USER]: user,
 });
